// src/app/models/sjcs/RetencionIrpfItem.ts
import { ComboItem } from '../../features/administracion/parametros/parametros-generales/parametros-generales.component';
export class RetencionIrpfItem {
  idRetencion: string;
  descripcion: string;
  idInstitucion: string;
  fechaModificacion: string;
  fechabaja: string;
  claveModelo: string;
  tipoSociedad: string;
  descripcionSociedad: string;
  retencion: string;
  retencionReal;
  historico: boolean;
  constructor() { }
}